Initialize Sentry in all processes and adopt current SDK conventions (#950 follow-up)
PR #960 revealed that instrumentation.ts had never loaded, which meant
server-side Sentry had never initialized. Fixing the load exposed the rest of
the gaps this commit closes:

- The worker process never called Sentry.init (init lived only in the
  sentry.*.config.ts files, which only Next instrumentation loads), so its
  Sentry.captureException calls for failed background jobs have always been
  silent no-ops. Extract the init into src/server/sentry.ts (initSentry) and
  call it from the worker and Discord bot at startup — this also installs
  Sentry's global uncaught-exception/unhandled-rejection handlers in those
  processes. sentry.server.config.ts and sentry.edge.config.ts now delegate
  to the same function.
- Export onRequestError = Sentry.captureRequestError from
  src/instrumentation.ts (the documented mechanism since Next 15 /
  @sentry/nextjs 8.28). Without it only explicit captureException call sites
  (like the tRPC error middleware) reach Sentry; now Server Components, plain
  route handlers (SSE, REST, OAuth, webhooks, MCP), and proxy errors are
  captured too.
- Flush buffered events on shutdown: Sentry.close(2000) in the app server's
  resource cleanup and in the worker and bot shutdown paths, so errors
  captured just before SIGTERM (including during the connection drain) are
  not dropped by process.exit.
- Replace deprecated sentry.client.config.ts with the
  src/instrumentation-client.ts convention (Next resolves it from src/ first,
  same as instrumentation.ts; the old file is not loaded at all under
  Turbopack). Register replayIntegration(), which the old config was missing
  — its replay sample rates did nothing without it. Export
  onRouterTransitionStart per convention.
- Plumb NEXT_PUBLIC_SENTRY_DSN as a Docker build ARG with a commented
  fly.toml [build.args] example: NEXT_PUBLIC_* vars are inlined at build
  time, so a runtime secret alone can never enable client Sentry.

Verified: instrumentation-client bundles into the client entry with the DSN
inlined (checked with a dummy build arg); onRequestError present in
.next/server/instrumentation.js; worker bundle boots, processes jobs, and
shuts down cleanly with and without a DSN (an invalid DSN degrades
gracefully); app-server graceful shutdown still completes with the flush in
the cleanup path; typecheck/lint/unit/e2e-against-production-build all pass.
